Add files via upload
Use this program to design your own chord progressions for archinovica. This is useful if you have a specific harmonic plan you want to follow in 5-limit space.  It will allow you to visualize your progression and easily calculate the necessary intonation pedaling to produce the desired result.  See README for more details.